+++ clang-tools-extra/clang-tidy/bugprone/NowFunctionCheck.cpp
@@ -1,0 +1,39 @@
+//===--- NowFunctionCheck.cpp - clang-tidy --------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "NowFunctionCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace bugprone {
+
+void NowFunctionCheck::registerMatchers(MatchFinder *Finder) {
+  Finder->addMatcher(callExpr().bind("y"), this);
+}
+
+void NowFunctionCheck::check(const MatchFinder::MatchResult &Result) {
+  const auto *MatchedCall = Result.Nodes.getNodeAs<CallExpr>("y");
+  if (MatchedCall) {
+    if (MatchedCall->getCalleeDecl()
+            ->getAsFunction()
+            ->getQualifiedNameAsString() == "rclcpp::now") {
+
+      diag(MatchedCall->getExprLoc(), "Dont call now()!")
+          << FixItHint::CreateRemoval(
+                 {MatchedCall->getBeginLoc(), MatchedCall->getEndLoc()});
+    }
+  }
+}
+
+} // namespace bugprone
+} // namespace tidy
+} // namespace clang