+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -99,6 +99,11 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`bugprone-now-function
+  <clang-tidy/checks/bugprone/now-function>` check.
+
+  FIXME: add release notes.
+
 New check aliases
 ^^^^^^^^^^^^^^^^^